Move CUDNN's binary dependency discovery to CUDNN.jl.
